split use signed integers for counters
Some of the remaining bugs will be easier to find if we can assert
when values < 0

(cherry picked from commit e144cb31b81e72a9543096f7ff9fdfb3a2498342)

+++ intervals.c
@@ -73,8 +73,8 @@ static struct timeval interval_start[INTERVAL_COUNT] = { 0 };
 static struct timeval ref_window_size;
 
 static struct {
-	unsigned int bytes;
-	unsigned int packets;
+	int64_t bytes;
+	int64_t packets;
 } totals;
 
 static void clear_table(int table_idx)
